refactor: move DocumentColorAdapter to new file
